LayoutTests/fast/events/touch/ios/click-event-after-pointerdown-prevent-default-on-ancestor-expected.txt

Tap me
Verifies that calling preventDefault() on a pointerdown event dispatched to an ancestor does not suppress the synthetic click event when tapping a descendant. To manually test, tap the button inside the orange box; both pointerdown and click should fire on the parent div.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointerdownFiredOnParent is true
PASS clickFiredOnParent is true
PASS successfullyParsed is true

TEST COMPLETE
